+++ common/src/main/java/net/mehvahdjukaar/every_compat/api/TabAddMode.java
@@ -1,0 +1,7 @@
+package net.mehvahdjukaar.every_compat.api;
+
+public enum TabAddMode {
+    AFTER_SAME_WOOD,
+    AFTER_SAME_TYPE,
+    AFTER_ALL
+}