+++ common/src/main/java/net/mehvahdjukaar/every_compat/modules/twigs/TwigsModule.java
@@ -1,13 +1,13 @@
 package net.mehvahdjukaar.every_compat.modules.twigs;
 
 import com.ninni.twigs.block.TableBlock;
+import net.mehvahdjukaar.every_compat.api.AbstractSimpleEntrySet;
 import net.mehvahdjukaar.every_compat.api.SimpleEntrySet;
 import net.mehvahdjukaar.every_compat.api.SimpleModule;
 import net.mehvahdjukaar.moonlight.api.set.wood.WoodType;
 import net.mehvahdjukaar.moonlight.api.set.wood.WoodTypeRegistry;
 import net.mehvahdjukaar.moonlight.api.util.Utils;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.core.Registry;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.world.level.block.Block;
@@ -26,6 +26,8 @@ public class TwigsModule extends SimpleModule {
                 .addTag(modRes("tables"), Registries.ITEM)
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
                 .defaultRecipe()
+                .setTabKey(modRes("twigs"))
+                .setTabMode(AbstractSimpleEntrySet.TabAddMode.AFTER_TYPE)
                 .setRenderType(() -> RenderType::cutout)
                 .createPaletteFromOak((p) -> p.remove(p.getDarkest()))
                 .addTexture(modRes("block/oak_table"))