+++ changelog.md
@@ -1,6 +1,6 @@
 ## FIXES:
-- Fixed the server not starting up correctly with **Woodworks**
-
+- Chipped module using tags in a non every compat namespace requiring restart to take effect
+- Creative tab config now does something woho!
 ---
 **LEGEND**:
 - (COMMON) = FORGE & FABRIC